refactor(SpaceAvatar) move SpaceAvatar.js to /SpaceList folder
