--- src/components/SpaceAvatar/README.md
@@ -1,145 +1,0 @@
-SpaceAvatar:
-
-```jsx
-<div>
-  <SpaceAvatar
-    id={1}
-    title="Valera Kotovski"
-    placeholder="empty"
-    size={40}
-    image="https://picsum.photos/200/200/?1"
-  />
-  <SpaceAvatar
-    id={2}
-    title="Hello World"
-    placeholder="lblue"
-    size={40}
-    image="https://picsum.photos/200/200/?2"
-  />
-  <SpaceAvatar
-    id={3}
-    title="Bad Timmy"
-    placeholder="blue"
-    size={40}
-    image="https://picsum.photos/200/200/?3"
-  />
-  <SpaceAvatar
-    id={4}
-    title="Ashley Simpson"
-    placeholder="purple"
-    size={40}
-    image="https://picsum.photos/200/200/?4"
-  />
-  <SpaceAvatar
-    id={5}
-    title="Ray Charles"
-    placeholder="red"
-    size={40}
-    image="https://picsum.photos/200/200/?5"
-  />
-  <SpaceAvatar
-    id={6}
-    title="Someone Else"
-    placeholder="orange"
-    size={40}
-    image="https://picsum.photos/200/200/?6"
-  />
-  <SpaceAvatar
-    id={7}
-    title="Net Vdohnoveniya"
-    placeholder="yellow"
-    size={40}
-    image="https://picsum.photos/200/200/?7"
-  />
-  <SpaceAvatar
-    id={8}
-    title="Vladimir Vladimirovich"
-    placeholder="green"
-    size={40}
-    image="https://picsum.photos/200/200/?8"
-  />
-</div>
-```
-
-SpaceAvatar without image:
-
-```jsx
-<div>
-  <SpaceAvatar
-    id={9}
-    title="Valera Kotovski"
-    placeholder="empty"
-    size={40}
-  />
-  <SpaceAvatar
-    id={10}
-    title="Hello World"
-    placeholder="lblue"
-    size={40}
-  />
-  <SpaceAvatar
-    id={11}
-    title="Bad Timmy"
-    placeholder="blue"
-    size={40}
-  />
-  <SpaceAvatar
-    id={12}
-    title="Ashley Simpson"
-    placeholder="purple"
-    size={40}
-  />
-  <SpaceAvatar
-    id={13}
-    title="Ray Charles"
-    placeholder="red"
-    size={40}
-  />
-  <SpaceAvatar
-    id={14}
-    title="s"
-    placeholder="orange"
-    size={40}
-  />
-  <SpaceAvatar
-    id={15}
-    title="W W"
-    placeholder="yellow"
-    size={40}
-  />
-  <SpaceAvatar
-    id={16}
-    title="Vladimir Vladimirovich"
-    placeholder="green"
-    size={40}
-  />
-</div>
-```
-
-SpaceAvatar with active state:
-
-```jsx
-<div>
-  <SpaceAvatar
-    id={17}
-    title="Valera Kotovski"
-    placeholder="empty"
-    active={true}
-    size={40}
-  />
-  <SpaceAvatar
-    id={18}
-    title="Hello World"
-    placeholder="lblue"
-    active={true}
-    size={40}
-  />
-  <SpaceAvatar
-    id={19}
-    title="Worst Wizard"
-    placeholder="blue"
-    active={true}
-    size={40}
-  />
-</div>
-```
