+++ src/components/SpaceList/SpaceAvatar.css
