+++ src/components/SpaceList/SpaceAvatar.js
